Implement the final solution for CCHAT lab
To solve the problem caused by the previous commit, the main idea
of this new solution is to create a new process for each of the
created channels, and make these new processes independent of the
server's process. This allows the server process to be killed, but
since the channels' processes are still up and running, the clients
in the same channel can still communicate with each other.